Restructure project into a monorepo with NPM workspaces
