+++ README.md
@@ -2,7 +2,25 @@
 
 ## Developers
 
+Clone the repository and install the dependencies:
+
+```bash
+$ git clone https://github.com/sleeyax/stremio-easynews-addon.git
+$ cd stremio-easynews-addon
+$ npm i
+```
+
+Run the easynews addon in development mode:
+
+```bash
+$ npm run start:addon:dev
+```
+
+## Production
+
 ```bash
 $ npm i
-$ npm run start:dev
+$ npm run build
+# to start the addon in production mode
+$ npm run start:addon
 ```
