+++ packages/addon/package.json
@@ -1,0 +1,22 @@
+{
+  "name": "@easynews/addon",
+  "version": "1.0.0",
+  "main": "dist/server.js",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1",
+    "build": "tsc",
+    "prepublish": "npm run build",
+    "start": "node dist/server.js",
+    "start:dev": "cross-env NODE_ENV=dev tsx watch src/server.ts",
+    "deploy": "beamup"
+  },
+  "description": "Provides content from Easynews and includes a search catalog.",
+  "devDependencies": {
+    "@types/stremio-addon-sdk": "^1.6.11",
+    "beamup-cli": "^1.3.0"
+  },
+  "dependencies": {
+    "@easynews/api": "^1.0.0",
+    "stremio-addon-sdk": "^1.6.10"
+  }
+}
