+++ packages/addon/src/manifest.ts
