+++ packages/addon/src/stremio-addon-sdk.d.ts
