+++ packages/api/package.json
@@ -1,0 +1,10 @@
+{
+  "name": "@easynews/api",
+  "version": "1.0.0",
+  "main": "./dist/index.js",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1",
+    "build": "tsc"
+  },
+  "description": "Easynews API library"
+}
