+++ packages/api/src/utils.ts
@@ -1,0 +1,5 @@
+export function createBasic(username: string, password: string) {
+  const userInfo = `${username}:${password}`;
+
+  return `Basic ${Buffer.from(userInfo).toString('base64')}`;
+}
