+++ tsconfig.json
@@ -1,8 +1,3 @@
 {
-  "extends": "@tsconfig/node22/tsconfig.json",
-  "compilerOptions": {
-    "outDir": "dist",
-    "resolveJsonModule": true
-  },
-  "include": ["src", "stremio-addon-sdk.d.ts"]
+  "references": [{ "path": "packages/addon" }, { "path": "packages/api" }]
 }
